@@ -22,7 +22,7 @@ if (NOT _CMAKE_BUILD_TYPE AND CMAKE_BUILD_TYPE)
 	unset ( CMAKE_BUILD_TYPE CACHE )
 endif ()
 
-SET ( CMAKE_CXX_STANDARD 14 )
+SET ( CMAKE_CXX_STANDARD 17 )
 SET ( MANTICORE_CMAKE_DIR "${MANTICORE_SOURCE_DIR}/cmake" )
 SET ( CMAKE_MODULE_PATH "${MANTICORE_SOURCE_DIR}/cmake" )
 SET ( MANTICORE_MODULE_PATH "${CMAKE_MODULE_PATH}" )
