@@ -6821,14 +6821,14 @@ static void QueryDiskChunks ( const CSphQuery & tQuery, CSphQueryResultMeta & tR
 	}
 
 	std::atomic<bool> bInterrupt {false};
-	auto fnCheckInterrupt = [&bInterrupt]() { return bInterrupt.load ( std::memory_order_relaxed ); };
+	auto CheckInterrupt = [&bInterrupt]() { return bInterrupt.load ( std::memory_order_relaxed ); };
 
 	Coro::ExecuteN ( tClonableCtx.Concurrency ( iJobs ), [&]
 	{
 		auto pSource = pDispatcher->MakeSource();
 		int iJob = -1; // make it consumed
 
-		if ( !pSource->FetchTask ( iJob ) || fnCheckInterrupt() )
+		if ( !pSource->FetchTask ( iJob ) || CheckInterrupt() )
 		{
 			sphLogDebug ( "Early finish parallel QueryDiskChunks because of empty queue" );
 			return; // already nothing to do, early finish.
@@ -6836,18 +6836,21 @@ static void QueryDiskChunks ( const CSphQuery & tQuery, CSphQueryResultMeta & tR
 
 		auto tJobContext = tClonableCtx.CloneNewContext ( !iJob );
 		auto& tCtx = tJobContext.first;
+		auto Interrupt = [&bInterrupt, &tCtx] ( const char* szReason ) {
+			tCtx.m_tMeta.m_sWarning = szReason;
+			bInterrupt.store ( true, std::memory_order_relaxed );
+		};
 		sphLogDebug ( "QueryDiskChunks cloned context %d (job %d)", tJobContext.second, iJob );
 		tClonableCtx.SetJobOrder ( tJobContext.second, -iJob ); // fixme! Same as in single search, but here we walk in reverse order. Need to fix?
-		Threads::Coro::Throttler_c tThrottler ( session::GetThrottlingPeriodMS () );
-		int iTick=1; // num of times coro rescheduled by throttler
-		while ( !fnCheckInterrupt() ) // some earlier job met error; abort.
+		Threads::Coro::SetThrottlingPeriod ( session::GetThrottlingPeriodMS() );
+		while ( !CheckInterrupt() ) // some earlier job met error; abort.
 		{
 			// jobs come in ascending order from 0 up to iJobs-1.
 			// We walk over disk chunk in reverse order, from last to 0-th.
 			auto iChunk = iJobs - iJob - 1;
 			sphLogDebug ( "QueryDiskChunks %d, Jb/Chunk: %d/%d", tJobContext.second, iJob, iChunk );
 			iJob = -1; // mark it consumed
-			myinfo::SetTaskInfo ( "%d ch %d:", iTick, iChunk );
+			myinfo::SetTaskInfo ( "%d ch %d:", Threads::Coro::NumOfRestarts(), iChunk );
 			auto & dLocalSorters = tCtx.m_dSorters;
 			CSphQueryResultMeta tChunkMeta;
 			CSphQueryResult tChunkResult;
@@ -6872,25 +6875,14 @@ static void QueryDiskChunks ( const CSphQuery & tQuery, CSphQueryResultMeta & tR
 
 			// check terms inconsistency among disk chunks
 			tThMeta.MergeWordStats ( tChunkMeta );
-
 			tThMeta.m_bHasPrediction |= tChunkMeta.m_bHasPrediction;
-
 			tSSTransform.Set ( iChunk, tChunkResult );
 
 			if ( tThMeta.m_bHasPrediction )
 				tThMeta.m_tStats.Add ( tChunkMeta.m_tStats );
 
 			if ( iChunk && sph::TimeExceeded ( tmMaxTimer ) )
-			{
-				tThMeta.m_sWarning = "query time exceeded max_query_time";
-				bInterrupt.store ( true, std::memory_order_relaxed );
-			}
-
-			if ( session::GetKilled() )
-			{
-				tThMeta.m_sWarning = "query was killed";
-				bInterrupt.store ( true, std::memory_order_relaxed );
-			}
+				Interrupt ( "query time exceeded max_query_time" );
 
 			if ( tThMeta.m_sWarning.IsEmpty() && !tChunkMeta.m_sWarning.IsEmpty() )
 				tThMeta.m_sWarning = tChunkMeta.m_sWarning;
@@ -6898,17 +6890,21 @@ static void QueryDiskChunks ( const CSphQuery & tQuery, CSphQueryResultMeta & tR
 			tThMeta.m_bTotalMatchesApprox |= tChunkMeta.m_bTotalMatchesApprox;
 			tThMeta.m_tIteratorStats.Merge ( tChunkMeta.m_tIteratorStats );
 
-			if ( fnCheckInterrupt() && !tChunkMeta.m_sError.IsEmpty() )
+			if ( CheckInterrupt() && !tChunkMeta.m_sError.IsEmpty() )
 				// FIXME? maybe handle this more gracefully (convert to a warning)?
 				tThMeta.m_sError = tChunkMeta.m_sError;
 
-			if ( !pSource->FetchTask ( iJob ) || fnCheckInterrupt() )
+			if ( !pSource->FetchTask ( iJob ) || CheckInterrupt() )
 				return; // all is done
 
 			// yield and reschedule every quant of time. It gives work to other tasks
-			if ( tThrottler.ThrottleAndKeepCrashQuery() )
-				// report current disk chunk processing
-				++iTick;
+			if ( Threads::Coro::RuntimeExceeded() )
+			{
+				if ( session::GetKilled() )
+					Interrupt ( "query was killed" );
+				else
+					Threads::Coro::RescheduleAndKeepCrashQuery();
+			}
 		}
 	});
 	sphLogDebug ( "QueryDiskChunks processed in %d thread(s)", tClonableCtx.NumWorked() );
@@ -7078,10 +7074,14 @@ static bool PerformFullscan ( const VecTraits_T<RtSegmentRefPtf_t> & dRamChunks,
 				return true;
 			}
 
-			if ( session::GetKilled() )
+			if ( Threads::Coro::RuntimeExceeded() )
 			{
-				sWarning = "query was killed";
-				return true;
+				if ( session::GetKilled() )
+				{
+					sWarning = "query was killed";
+					return true;
+				}
+				Threads::Coro::RescheduleAndKeepCrashQuery();
 			}
 		}
 	}
