@@ -31,7 +31,7 @@ inline int sphBitCount ( uint64_t n )
 // fix last line for 64-bit numbers
 inline int sphBitCount ( DWORD n )
 {
-	register DWORD tmp;
+	DWORD tmp;
 	tmp = n - ( ( n >> 1 ) & 033333333333 ) - ( ( n >> 2 ) & 011111111111 );
 	return ( ( tmp + ( tmp >> 3 ) ) & 030707070707 ) % 63;
 }
